+++ src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './env-adapter';
+export { wagmiConfig, targetChain } from './wagmi-config';
