+++ src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './auth';
+export * from './deposit';
