+++ src/store/rpc-thunks.ts
@@ -1,0 +1,65 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { getRpc, queryConfig as queryConfigService, queryState as queryStateService } from '../rpc/client';
+
+export const getConfig = createAsyncThunk(
+  'client/getConfig',
+  async () => {
+    const res: any = await queryConfigService();
+    const data = JSON.parse(res.data);
+    return data;
+  }
+);
+
+export const sendTransaction = createAsyncThunk(
+  'client/sendTransaction',
+  async (params: { cmd: BigUint64Array, prikey: string }, { rejectWithValue }) => {
+    try {
+      const { cmd, prikey } = params;
+      const state: any = await getRpc().sendTransaction(cmd, prikey);
+      console.log("(Data-Transaction)", state);
+      return state;
+    } catch (err: any) {
+      return rejectWithValue(err);
+    }
+  }
+);
+
+export const sendExtrinsicTransaction = createAsyncThunk(
+  'client/sendExtrinsicTransaction',
+  async (params: { cmd: BigUint64Array, prikey: string }, { rejectWithValue }) => {
+    try {
+      const { cmd, prikey } = params;
+      const state: any = await getRpc().sendExtrinsic(cmd, prikey);
+      console.log("(Data-Transaction)", state);
+      return state;
+    } catch (err: any) {
+      return rejectWithValue(err);
+    }
+  }
+);
+
+export const queryState = createAsyncThunk(
+  'client/queryState',
+  async (key: string, { rejectWithValue }) => {
+    try {
+      const state: any = await queryStateService(key);
+      console.log("(Data-QueryState)", state);
+      return state;
+    } catch (err: any) {
+      return rejectWithValue(err);
+    }
+  }
+);
+
+export const queryInitialState = createAsyncThunk(
+  'client/queryInitialState',
+  async (key: string, { rejectWithValue }) => {
+    try {
+      const state: any = await queryStateService(key);
+      console.log("(Data-QueryState)", state);
+      return state;
+    } catch (err: any) {
+      return rejectWithValue(err);
+    }
+  }
+);
