+++ src/store/selectors.ts
@@ -1,0 +1,5 @@
+import { State } from '../types';
+
+export const selectL1Account = <T extends State>(state: T) => state.account.l1Account;
+export const selectL2Account = <T extends State>(state: T) => state.account.l2account;
+export const selectLoginStatus = <T extends State>(state: T) => state.account.status;
